rebase, move class & fix test
