+++ src/main/java/com/github/fppt/jedismock/operations/keys/Persist.java
@@ -1,17 +1,20 @@
-package com.github.fppt.jedismock.operations;
+package com.github.fppt.jedismock.operations.keys;
 
 import com.github.fppt.jedismock.datastructures.Slice;
+import com.github.fppt.jedismock.operations.AbstractRedisOperation;
+import com.github.fppt.jedismock.operations.RedisCommand;
 import com.github.fppt.jedismock.server.Response;
 import com.github.fppt.jedismock.storage.RedisBase;
+
 import java.util.List;
 
 @RedisCommand("persist")
-class RO_persist extends AbstractRedisOperation {
-    RO_persist(RedisBase base, List<Slice> params) {
+class Persist extends AbstractRedisOperation {
+    Persist(RedisBase base, List<Slice> params) {
         super(base, params);
     }
 
-    Slice response() {
+    protected Slice response() {
         return Response.integer(base().setDeadline(params().get(0), -1));
     }
 }
